# fix performance regression in all releases-views