@syphar 2021-01-02
I'm not sure which way to go now, probably a decision for the maintainers. 

Easiest approach is reducing the results for the stars-views.