@Nemo157 2021-01-04
Can we make `releases.release_time` non-nullable? We appear to just default it to `now` if we fail to get it from crates.io.