@Nemo157 2021-01-04
It actually is already `NOT NULL`; the `NULLS LAST` is only applied when sorting by stars.