@syphar 2021-01-04
Nulls last is also used for the release time, 

Also some other indexes were missing