@syphar 2021-01-04
for example one on `releases.github_repo` which is used for the join to `github_repos.id`

I'll finish everything up, then you'll see